feat: move interfaces and types to types.ts
