+++ .env.local
@@ -1,0 +1,1 @@
+SPACEX_URL=https://api.spacex.land/graphql/
