+++ package.json
@@ -116,7 +116,7 @@
     "react-intl": "5.20.3",
     "react-redux": "7.0.2",
     "react-router-dom": "^5.1.2",
-    "react-screentype-hook": "^1.0.15",
+    "react-screentype-hook": "^1.0.18",
     "redux-injectors": "^1.3.0",
     "redux-persist": "^6.0.0",
     "redux-persist-transform-immutable": "^5.0.0",
